feat: #16 JPA 추가
JPA 추가

close #16
