+++ src/main/java/com/example/fastcampusmysql/domain/follow/entity/Follow.java
@@ -2,16 +2,24 @@ package com.example.fastcampusmysql.domain.follow.entity;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
 @Getter
+@Entity
+@NoArgsConstructor
 public class Follow {
-    private final Long id;
-    private final Long fromMemberId;
-    private final Long toMemberId;
-    private final LocalDateTime createdAt;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private Long fromMemberId;
+    private Long toMemberId;
+    private LocalDateTime createdAt;
 
     @Builder
     public Follow(Long id, Long fromMemberId, Long toMemberId, LocalDateTime createdAt) {
