+++ src/main/java/com/example/fastcampusmysql/domain/follow/repository/FollowRepository.java
@@ -17,49 +17,8 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-@Repository
-@RequiredArgsConstructor
-public class FollowRepository {
-    final private NamedParameterJdbcTemplate namedParameterJdbcTemplate;
-    private static final String TABLE = "Follow";
-    private static final RowMapper<Follow> rowMapper = (ResultSet resultSet, int rowNum) -> Follow
-            .builder()
-            .id(resultSet.getLong("id"))
-            .fromMemberId(resultSet.getLong("fromMemberId"))
-            .toMemberId(resultSet.getLong("toMemberId"))
-            .createdAt(resultSet.getObject("createdAt", LocalDateTime.class))
-            .build();
-
-    public List<Follow> findAllbyFromMemberId(Long fromMemberId) {
-        String sql = String.format("SELECT * FROM %s WHERE fromMemberId = :fromMemberId", TABLE);
-        SqlParameterSource params = new MapSqlParameterSource().addValue("fromMemberId", fromMemberId);
-        return namedParameterJdbcTemplate.query(sql, params, rowMapper);
-    }
-    public Follow save(Follow follow) {
-        if (follow.getId() == null)
-            return insert(follow);
-        throw new UnsupportedOperationException("Follow는 갱신을 지원하지 않습니다");
-    }
-
-    private Follow insert(Follow follow) {
-        SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(namedParameterJdbcTemplate.getJdbcTemplate())
-                .withTableName(TABLE)
-                .usingGeneratedKeyColumns("id");
-        SqlParameterSource params = new BeanPropertySqlParameterSource(follow);
-        Long id = simpleJdbcInsert.executeAndReturnKey(params).longValue();
-        return Follow
-                .builder()
-                .id(id)
-                .fromMemberId(follow.getFromMemberId())
-                .toMemberId(follow.getToMemberId())
-                .createdAt(follow.getCreatedAt())
-                .build();
-    }
-
-
-    public List<Follow> findAllbyToMemberId(Long toMemberId) {
-        String sql = String.format("SELECT * FROM %s WHERE toMemberId = :toMemberId", TABLE);
-        SqlParameterSource params = new MapSqlParameterSource().addValue("toMemberId", toMemberId);
-        return namedParameterJdbcTemplate.query(sql, params, rowMapper);
-    }
+public interface FollowRepository {
+    List<Follow> findAllByFromMemberId(Long fromMemberId);
+    Follow save(Follow follow);
+    List<Follow> findAllByToMemberId(Long toMemberId);
 }
