+++ src/main/java/com/example/fastcampusmysql/domain/follow/repository/JdbcFollowRepository.java
@@ -1,0 +1,62 @@
+package com.example.fastcampusmysql.domain.follow.repository;
+
+import com.example.fastcampusmysql.domain.follow.entity.Follow;
+import lombok.RequiredArgsConstructor;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+
+import java.sql.ResultSet;
+import java.time.LocalDateTime;
+import java.util.List;
+
+//@Repository
+@RequiredArgsConstructor
+public class JdbcFollowRepository implements FollowRepository {
+    final private NamedParameterJdbcTemplate namedParameterJdbcTemplate;
+    private static final String TABLE = "Follow";
+    private static final RowMapper<Follow> rowMapper = (ResultSet resultSet, int rowNum) -> Follow
+            .builder()
+            .id(resultSet.getLong("id"))
+            .fromMemberId(resultSet.getLong("fromMemberId"))
+            .toMemberId(resultSet.getLong("toMemberId"))
+            .createdAt(resultSet.getObject("createdAt", LocalDateTime.class))
+            .build();
+
+    public List<Follow> findAllByFromMemberId(Long fromMemberId) {
+        String sql = String.format("SELECT * FROM %s WHERE fromMemberId = :fromMemberId", TABLE);
+        SqlParameterSource params = new MapSqlParameterSource().addValue("fromMemberId", fromMemberId);
+        return namedParameterJdbcTemplate.query(sql, params, rowMapper);
+    }
+    public Follow save(Follow follow) {
+        if (follow.getId() == null)
+            return insert(follow);
+        throw new UnsupportedOperationException("Follow는 갱신을 지원하지 않습니다");
+    }
+
+    private Follow insert(Follow follow) {
+        SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(namedParameterJdbcTemplate.getJdbcTemplate())
+                .withTableName(TABLE)
+                .usingGeneratedKeyColumns("id");
+        SqlParameterSource params = new BeanPropertySqlParameterSource(follow);
+        Long id = simpleJdbcInsert.executeAndReturnKey(params).longValue();
+        return Follow
+                .builder()
+                .id(id)
+                .fromMemberId(follow.getFromMemberId())
+                .toMemberId(follow.getToMemberId())
+                .createdAt(follow.getCreatedAt())
+                .build();
+    }
+
+
+    public List<Follow> findAllByToMemberId(Long toMemberId) {
+        String sql = String.format("SELECT * FROM %s WHERE toMemberId = :toMemberId", TABLE);
+        SqlParameterSource params = new MapSqlParameterSource().addValue("toMemberId", toMemberId);
+        return namedParameterJdbcTemplate.query(sql, params, rowMapper);
+    }
+}
