+++ src/main/java/com/example/fastcampusmysql/domain/follow/service/FollowReadService.java
@@ -13,9 +13,9 @@ public class FollowReadService {
     private final FollowRepository followRepository;
 
     public List<Follow> getFollowings(Long memberId) {
-        return followRepository.findAllbyFromMemberId(memberId);
+        return followRepository.findAllByFromMemberId(memberId);
     }
     public List<Follow> getFollowers(Long memberId) {
-        return followRepository.findAllbyToMemberId(memberId);
+        return followRepository.findAllByToMemberId(memberId);
     }
 }
